# Avoid dynamic dispatch when calling wrapCheck